Split config_setup.h file
