+++ src/config/config_definition.cc
@@ -34,13 +34,17 @@
 #include "config/result/box_layout.h"
 #include "config/result/client_config.h"
 #include "config/result/transcoding.h"
+#include "config/setup/config_setup_array.h"
 #include "config/setup/config_setup_autoscan.h"
+#include "config/setup/config_setup_bool.h"
 #include "config/setup/config_setup_boxlayout.h"
 #include "config/setup/config_setup_client.h"
 #include "config/setup/config_setup_dictionary.h"
 #include "config/setup/config_setup_dynamic.h"
 #include "config/setup/config_setup_enum.h"
+#include "config/setup/config_setup_int.h"
 #include "config/setup/config_setup_path.h"
+#include "config/setup/config_setup_string.h"
 #include "config/setup/config_setup_time.h"
 #include "config/setup/config_setup_transcoding.h"
 #include "config/setup/config_setup_tweak.h"
