+++ src/config/config_setup.h
@@ -44,13 +44,6 @@
 #define ITEM_PATH_PREFIX (-3)
 
 using StringCheckFunction = std::function<bool(std::string& value)>;
-using ArrayInitFunction = std::function<bool(const pugi::xml_node& value, std::vector<std::string>& result, const char* node_name)>;
-using ArrayItemCheckFunction = std::function<bool(const std::string& value)>;
-using IntCheckFunction = std::function<bool(int value)>;
-using IntParseFunction = std::function<int(const std::string& value)>;
-using IntPrintFunction = std::function<std::string(int value)>;
-using IntMinFunction = std::function<bool(int value, int minValue)>;
-
 using ConfigOptionIterator = EnumIterator<config_option_t, config_option_t::CFG_MIN, config_option_t::CFG_MAX>;
 
 class ConfigValue {
@@ -169,244 +162,6 @@ public:
 
     virtual std::string getUniquePath() const { return xpath; }
     virtual std::string getCurrentValue() const { return optionValue ? optionValue->getOption() : ""; }
-};
-
-class ConfigStringSetup : public ConfigSetup {
-protected:
-    bool notEmpty = false;
-
-public:
-    ConfigStringSetup(config_option_t option, const char* xpath, const char* help, bool required = false, const char* defaultValue = "", bool notEmpty = false)
-        : ConfigSetup(option, xpath, help, required, defaultValue)
-        , notEmpty(notEmpty)
-    {
-    }
-
-    ConfigStringSetup(config_option_t option, const char* xpath, const char* help, const char* defaultValue, StringCheckFunction check = nullptr, bool notEmpty = false, bool required = false)
-        : ConfigSetup(option, xpath, help, std::move(check), defaultValue, required)
-        , notEmpty(notEmpty)
-    {
-    }
-
-    std::string getTypeString() const override { return "String"; }
-
-    void makeOption(const pugi::xml_node& root, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments = nullptr) override;
-
-    std::shared_ptr<ConfigOption> newOption(const std::string& optValue);
-
-    static bool CheckCharset(std::string& value);
-};
-
-class ConfigIntSetup : public ConfigSetup {
-protected:
-    IntCheckFunction valueCheck = nullptr;
-    IntParseFunction parseValue = nullptr;
-    IntPrintFunction printValue = nullptr;
-    IntMinFunction minCheck = nullptr;
-    int minValue {};
-
-public:
-    ConfigIntSetup(config_option_t option, const char* xpath, const char* help)
-        : ConfigSetup(option, xpath, help)
-    {
-        this->defaultValue = fmt::to_string(0);
-    }
-
-    ConfigIntSetup(config_option_t option, const char* xpath, const char* help, int defaultValue)
-        : ConfigSetup(option, xpath, help)
-    {
-        this->defaultValue = fmt::to_string(defaultValue);
-    }
-
-    ConfigIntSetup(config_option_t option, const char* xpath, const char* help, IntCheckFunction check)
-        : ConfigSetup(option, xpath, help)
-        , valueCheck(std::move(check))
-    {
-        this->defaultValue = fmt::to_string(0);
-    }
-
-    ConfigIntSetup(config_option_t option, const char* xpath, const char* help, int defaultValue, IntParseFunction parseValue, IntPrintFunction printValue = nullptr)
-        : ConfigSetup(option, xpath, help)
-        , parseValue(std::move(parseValue))
-        , printValue(std::move(printValue))
-    {
-        this->defaultValue = fmt::to_string(defaultValue);
-    }
-
-    ConfigIntSetup(config_option_t option, const char* xpath, const char* help, int defaultValue, IntCheckFunction check, IntPrintFunction printValue = nullptr)
-        : ConfigSetup(option, xpath, help)
-        , valueCheck(std::move(check))
-        , printValue(std::move(printValue))
-    {
-        this->defaultValue = fmt::to_string(defaultValue);
-    }
-
-    ConfigIntSetup(config_option_t option, const char* xpath, const char* help, int defaultValue, int minValue, IntMinFunction check)
-        : ConfigSetup(option, xpath, help)
-        , minCheck(std::move(check))
-        , minValue(minValue)
-    {
-        this->defaultValue = fmt::to_string(defaultValue);
-    }
-
-    ConfigIntSetup(config_option_t option, const char* xpath, const char* help, const char* defaultValue, IntCheckFunction check = nullptr, IntPrintFunction printValue = nullptr)
-        : ConfigSetup(option, xpath, help)
-        , valueCheck(std::move(check))
-        , printValue(std::move(printValue))
-    {
-        this->defaultValue = defaultValue;
-    }
-
-    ConfigIntSetup(config_option_t option, const char* xpath, const char* help, const char* defaultValue, StringCheckFunction check, IntPrintFunction printValue = nullptr)
-        : ConfigSetup(option, xpath, help, std::move(check), defaultValue)
-        , printValue(std::move(printValue))
-    {
-    }
-
-    std::string getTypeString() const override { return parseValue ? "String" : "Number"; }
-
-    void makeOption(const pugi::xml_node& root, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments = nullptr) override;
-
-    void makeOption(std::string optValue, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments = nullptr) override;
-
-    int getXmlContent(const pugi::xml_node& root);
-
-    std::shared_ptr<ConfigOption> newOption(int optValue);
-
-    int checkIntValue(std::string& sVal, const std::string& pathName = "") const;
-
-    std::string getCurrentValue() const override { return optionValue ? optionValue->getOption() : ""; }
-
-    static bool CheckProfileNumberValue(std::string& value);
-
-    static bool CheckMinValue(int value, int minValue);
-
-    static bool CheckImageQualityValue(int value);
-
-    static bool CheckPortValue(int value);
-
-    static bool CheckUpnpStringLimitValue(int value);
-};
-
-class ConfigBoolSetup : public ConfigSetup {
-    using ConfigSetup::ConfigSetup;
-
-public:
-    ConfigBoolSetup(config_option_t option, const char* xpath, const char* help, const char* defaultValue, StringCheckFunction check = nullptr)
-        : ConfigSetup(option, xpath, help, std::move(check), defaultValue)
-    {
-    }
-
-    ConfigBoolSetup(config_option_t option, const char* xpath, const char* help, bool defaultValue, StringCheckFunction check = nullptr)
-        : ConfigSetup(option, xpath, help, std::move(check))
-    {
-        this->defaultValue = defaultValue ? YES : NO;
-    }
-
-    ConfigBoolSetup(config_option_t option, const char* xpath, const char* help, bool defaultValue, bool required)
-        : ConfigSetup(option, xpath, help, required)
-    {
-        this->defaultValue = defaultValue ? YES : NO;
-    }
-
-    std::string getTypeString() const override { return "Boolean"; }
-
-    void makeOption(const pugi::xml_node& root, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments = nullptr) override;
-
-    void makeOption(std::string optValue, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments = nullptr) override;
-
-    bool getXmlContent(const pugi::xml_node& root);
-
-    bool checkValue(std::string& optValue, const std::string& pathName = "") const;
-
-    std::shared_ptr<ConfigOption> newOption(bool optValue);
-
-    std::string getCurrentValue() const override { return optionValue ? fmt::to_string(optionValue->getBoolOption()) : ""; }
-
-    static bool CheckInotifyValue(std::string& value);
-
-    static bool CheckMarkPlayedValue(std::string& value);
-};
-
-class ConfigArraySetup : public ConfigSetup {
-protected:
-    bool notEmpty = false;
-    bool itemNotEmpty = false;
-    ArrayInitFunction initArray = nullptr;
-    ArrayItemCheckFunction itemCheck = nullptr;
-    std::vector<std::string> defaultEntries;
-
-    /// \brief Creates an array of strings from an XML nodeset.
-    /// \param element starting element of the nodeset.
-    /// \param nodeOption name of each node in the set
-    /// \param attrOption name of the attribute, the value of which shouldbe extracted
-    ///
-    /// Similar to \fn createDictionaryFromNode() this one extracts
-    /// data from the following XML:
-    /// <some-section>
-    ///     <tag attr="data"/>
-    ///     <tag attr="otherdata"/>
-    /// <some-section>
-    ///
-    /// This function will create an array like that: ["data", "otherdata"]
-    bool createOptionFromNode(const pugi::xml_node& element, std::vector<std::string>& result) const;
-
-    bool updateItem(std::size_t i, const std::string& optItem, const std::shared_ptr<Config>& config, const std::shared_ptr<ArrayOption>& value, const std::string& optValue, const std::string& status = "") const;
-
-public:
-    config_option_t nodeOption;
-    config_option_t attrOption = CFG_MAX;
-
-    ConfigArraySetup(config_option_t option, const char* xpath, const char* help, config_option_t nodeOption,
-        ArrayInitFunction init = nullptr, bool notEmpty = false, std::vector<std::string> defaultEntries = {})
-        : ConfigSetup(option, xpath, help)
-        , notEmpty(notEmpty)
-        , initArray(std::move(init))
-        , defaultEntries(std::move(defaultEntries))
-        , nodeOption(nodeOption)
-    {
-    }
-
-    ConfigArraySetup(config_option_t option, const char* xpath, const char* help, config_option_t nodeOption, config_option_t attrOption,
-        bool notEmpty, bool itemNotEmpty, std::vector<std::string> defaultEntries = {})
-        : ConfigSetup(option, xpath, help)
-        , notEmpty(notEmpty)
-        , itemNotEmpty(itemNotEmpty)
-        , defaultEntries(std::move(defaultEntries))
-        , nodeOption(nodeOption)
-        , attrOption(attrOption)
-    {
-    }
-
-    ConfigArraySetup(config_option_t option, const char* xpath, const char* help, config_option_t nodeOption, config_option_t attrOption,
-        ArrayItemCheckFunction itemCheck, std::vector<std::string> defaultEntries = {})
-        : ConfigSetup(option, xpath, help)
-        , itemCheck(std::move(itemCheck))
-        , defaultEntries(std::move(defaultEntries))
-        , nodeOption(nodeOption)
-        , attrOption(attrOption)
-    {
-    }
-
-    std::string getTypeString() const override { return "List"; }
-
-    void makeOption(const pugi::xml_node& root, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments = nullptr) override;
-
-    bool updateDetail(const std::string& optItem, std::string& optValue, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments = nullptr) override;
-
-    std::string getItemPath(int index = 0, config_option_t propOption = CFG_MAX, config_option_t propOption2 = CFG_MAX, config_option_t propOption3 = CFG_MAX, config_option_t propOption4 = CFG_MAX) const override;
-
-    std::vector<std::string> getXmlContent(const pugi::xml_node& optValue);
-
-    bool checkArrayValue(const std::string& value, std::vector<std::string>& result) const;
-
-    std::shared_ptr<ConfigOption> newOption(const std::vector<std::string>& optValue);
-
-    std::string getCurrentValue() const override { return {}; }
-
-    static bool InitPlayedItemsMark(const pugi::xml_node& value, std::vector<std::string>& result, const char* nodeName);
-
-    static bool InitItemsPerPage(const pugi::xml_node& value, std::vector<std::string>& result, const char* nodeName);
 };
 
 #endif // __CONFIG_SETUP_H__
