+++ src/config/setup/config_setup_array.cc
@@ -22,12 +22,13 @@
 
 /// \file config_setup_array.cc
 
-#include "config/config_setup.h" // API
+#include "config_setup_array.h" // API
 
 #include <numeric>
 
 #include "config/config_definition.h"
 #include "config/config_options.h"
+#include "config_setup_int.h"
 
 /// \brief Creates an array of strings from an XML nodeset.
 /// \param element starting element of the nodeset.
