+++ src/config/setup/config_setup_autoscan.cc
@@ -30,9 +30,12 @@
 #include "config/config_option_enum.h"
 #include "config/config_options.h"
 #include "config/result/autoscan.h"
-#include "config/setup/config_setup_enum.h"
-#include "config/setup/config_setup_path.h"
-#include "config/setup/config_setup_time.h"
+#include "config_setup_bool.h"
+#include "config_setup_enum.h"
+#include "config_setup_int.h"
+#include "config_setup_path.h"
+#include "config_setup_string.h"
+#include "config_setup_time.h"
 #include "content/autoscan_list.h"
 
 #ifdef HAVE_INOTIFY
