+++ src/config/setup/config_setup_client.cc
@@ -29,7 +29,10 @@
 #include "config/config_definition.h"
 #include "config/config_options.h"
 #include "config/result/client_config.h"
+#include "config_setup_bool.h"
 #include "config_setup_dictionary.h"
+#include "config_setup_int.h"
+#include "config_setup_string.h"
 
 /// \brief Creates an array of ClientConfig objects from a XML nodeset.
 /// \param element starting element of the nodeset.
