+++ src/config/setup/config_setup_transcoding.cc
@@ -30,9 +30,13 @@
 #include "config/config_option_enum.h"
 #include "config/config_options.h"
 #include "config/result/transcoding.h"
-#include "config/setup/config_setup_dictionary.h"
-#include "config/setup/config_setup_enum.h"
-#include "config/setup/config_setup_path.h"
+#include "config_setup_array.h"
+#include "config_setup_bool.h"
+#include "config_setup_dictionary.h"
+#include "config_setup_enum.h"
+#include "config_setup_int.h"
+#include "config_setup_path.h"
+#include "config_setup_string.h"
 #include "metadata/metadata_handler.h"
 #include "metadata/resolution.h"
