+++ src/config/setup/config_setup_tweak.cc
@@ -29,7 +29,9 @@
 #include "config/config_definition.h"
 #include "config/config_options.h"
 #include "config/result/directory_tweak.h"
-#include "config/setup/config_setup_path.h"
+#include "config_setup_bool.h"
+#include "config_setup_path.h"
+#include "config_setup_string.h"
 
 /// \brief Creates an array of DirectoryTweak objects from a XML nodeset.
 /// \param element starting element of the nodeset.
